text-translator-app/requirements.txt
streamlit
python-dotenv
langchain
langchain_community
openai